config: move config.go out of config/
closes #423
