+++ pkg/wallet/wallet_config.go
@@ -1,0 +1,7 @@
+package wallet
+
+// Config is a wallet info.
+type Config struct {
+	Path     string `yaml:"Path"`
+	Password string `yaml:"Password"`
+}
